refactor: extract text manipulation utilities from SearchView
- Create new textUtils.ts with normalizeWhitespace, stripMediaUrls, stripPreviewUrls, stripAllUrls
- Remove duplicate text manipulation functions from SearchView.tsx
- Update SearchView to use extracted utilities
- Reduce SearchView.tsx from 1626 to 1594 lines (-32 lines)
- Improve code organization and reusability
